+++ decidim-conferences/lib/decidim/conferences/admin_engine.rb
@@ -14,7 +14,7 @@ module Decidim
       paths["lib/tasks"] = nil
 
       routes do
-        resources :conferences, param: :slug, except: :show do
+        resources :conferences, param: :slug, except: [:show, :destroy] do
           resource :publish, controller: "conference_publications", only: [:create, :destroy]
           resources :copies, controller: "conference_copies", only: [:new, :create]
           resources :speakers, controller: "conference_speakers"